use wahltag with nummer and move WahltagWithNummer to common; its use by multiple services
