+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/services/common/WahltagWithNummer.java
@@ -1,4 +1,4 @@
-package de.muenchen.oss.wahllokalsystem.basisdatenservice.services.wahlen;
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.services.common;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
