+++ wls-basisdaten-service/src/test/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/services/wahltermindaten/AsyncWahltermindatenServiceTest.java
@@ -10,6 +10,7 @@ import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlen.Wahlart;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlvorschlag.Wahlvorschlaege;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.domain.wahlvorschlag.WahlvorschlaegeRepository;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.common.WahlbezirkArtModel;
+import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.common.WahltagWithNummer;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.kopfdaten.BasisdatenModel;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlagen.ReferendumvorlagenClient;
 import de.muenchen.oss.wahllokalsystem.basisdatenservice.services.referendumvorlagen.ReferendumvorlagenModel;
@@ -64,7 +65,7 @@ class AsyncWahltermindatenServiceTest {
             val wahltagNummer = "wahltagNummer";
             val basisdatenModel = createEmptyBasisdatenModel();
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(asyncProgress).reset(wahltagDate, wahltagNummer);
         }
@@ -81,7 +82,7 @@ class AsyncWahltermindatenServiceTest {
             Mockito.when(wahlvorschlaegeClient.getWahlvorschlaege(any())).thenReturn(mockedWahlvorschlaegeClientResponse);
             Mockito.when(wahlvorschlaegeModelMapper.toEntity(mockedWahlvorschlaegeClientResponse)).thenReturn(mockedWahlvorschlaegeModelMappedAsEntity);
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(wahlvorschlaegeRepository, times(4)).save(mockedWahlvorschlaegeModelMappedAsEntity);
             Mockito.verify(asyncProgress, times(4)).incWahlvorschlaegeFinished();
@@ -101,7 +102,7 @@ class AsyncWahltermindatenServiceTest {
             Mockito.when(wahlvorschlaegeModelMapper.toEntity(mockedWahlvorschlaegeClientResponse)).thenReturn(mockedWahlvorschlaegeModelMappedAsEntity);
             Mockito.doThrow(new RuntimeException("saving failed")).when(wahlvorschlaegeRepository).save(mockedWahlvorschlaegeModelMappedAsEntity);
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(asyncProgress, times(4)).incWahlvorschlaegeFinished();
             Mockito.verify(asyncProgress, times(4)).setWahlvorschlaegeNext(any());
@@ -118,7 +119,7 @@ class AsyncWahltermindatenServiceTest {
 
             Mockito.doThrow(new RuntimeException("getting data from client failed")).when(wahlvorschlaegeClient).getWahlvorschlaege(any());
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(asyncProgress, times(4)).incWahlvorschlaegeFinished();
             Mockito.verify(asyncProgress, times(4)).setWahlvorschlaegeNext(any());
@@ -137,7 +138,7 @@ class AsyncWahltermindatenServiceTest {
             Mockito.when(referendumvorlagenModelMapper.toEntity(eq(mockedReferendumvorlagenClientResponse), any()))
                     .thenReturn(mockedReferendumvorlagenModelMappedAsEntity);
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(referendumvorlagenRepository, times(4)).save(mockedReferendumvorlagenModelMappedAsEntity);
             Mockito.verify(asyncProgress, times(4)).incReferendumVorlagenFinished();
@@ -158,7 +159,7 @@ class AsyncWahltermindatenServiceTest {
                     .thenReturn(mockedReferendumvorlagenModelMappedAsEntity);
             Mockito.doThrow(new RuntimeException("saving failed")).when(referendumvorlagenRepository).save(any());
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(asyncProgress, times(4)).incReferendumVorlagenFinished();
             Mockito.verify(asyncProgress, times(4)).setReferendumVorlagenNext(any());
@@ -172,7 +173,7 @@ class AsyncWahltermindatenServiceTest {
 
             Mockito.doThrow(new RuntimeException("getting data from client failed")).when(referendumvorlagenClient).getReferendumvorlagen(any());
 
-            unitUnderTest.initVorlagenAndVorschlaege(wahltagDate, wahltagNummer, basisdatenModel);
+            unitUnderTest.initVorlagenAndVorschlaege(new WahltagWithNummer(wahltagDate, wahltagNummer), basisdatenModel);
 
             Mockito.verify(asyncProgress, times(4)).incReferendumVorlagenFinished();
             Mockito.verify(asyncProgress, times(4)).setReferendumVorlagenNext(any());
